Final commit for MS-1 User Registration Page
created a createDB.sql file for needed scripts to create tables needed for this milestone.
refactored the userRegistrationPage to flow with the schema for the user_info table.
removed role items as they're un needed for this deliverable.
refactored createUser.php to add logic for required fields as well as the needed logic for nickName per requirements.
Added throw sql conn error to creation of user for better troubleshooting.
removed un needed createDB.php
added folder for Project Reports.